Revert "Use big int for user id"
This reverts commit 8873c84f256d62e8b853ce83d9b6b532f5d0a1ce.

+++ roller_bot/models/roll.py
@@ -1,6 +1,6 @@
 from typing import List
 
-from sqlalchemy import Boolean, Column, DateTime, ForeignKey, Integer, BigInteger
+from sqlalchemy import Boolean, Column, DateTime, ForeignKey, Integer
 from sqlalchemy.orm import relationship
 
 from roller_bot.models.base import Base
@@ -10,7 +10,7 @@ from roller_bot.models.bonus_value import BonusValue
 class Roll(Base):
     __tablename__: str = "rolls"
     id: Column = Column(Integer, primary_key=True, autoincrement=True, nullable=False)
-    user_id: Column = Column(BigInteger, ForeignKey("users.id"), nullable=False)
+    user_id: Column = Column(Integer, ForeignKey("users.id"), nullable=False)
     item_id: Column = Column(Integer, ForeignKey("items.id"), nullable=False)
     roll_time: Column = Column(DateTime, nullable=False)
     base_value: Column = Column(Integer, nullable=False)
